carshopping/requirements.txt
beautifulsoup4==4.12.2
pandas==2.0.0
price-parser==0.3.4
requests==2.28.2
plotly.express
